requirements.txt
streamlit
audio-recorder-streamlit
azure-cognitiveservices-speech
openai
azure-cognitiveservices-vision-computervision
pillow

